% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-se.R, R/group.R, R/select-helpers.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{group_rows}
\alias{add_tally}
\alias{n}
\alias{everything}
\alias{all_of}
\alias{any_of}
\alias{one_of}
\alias{contains}
\alias{matches}
\alias{starts_with}
\alias{ends_with}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:count]{add_tally}}, \code{\link[dplyr:count]{add_tally}}, \code{\link[dplyr:group_data]{group_rows}}, \code{\link[dplyr:group_data]{group_rows}}, \code{\link[dplyr:context]{n}}}

  \item{tidyselect}{\code{\link[tidyselect]{all_of}}, \code{\link[tidyselect:all_of]{any_of}}, \code{\link[tidyselect:starts_with]{contains}}, \code{\link[tidyselect:starts_with]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect:starts_with]{matches}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{starts_with}}}
}}

